+++ zgw_consumers/legacy/__init__.py
@@ -1,0 +1,3 @@
+"""This module contains stuff from the legacy client.
+For newer versions, please make use of the ape-pie client instead.
+"""